Remake database handling structure
Separate tables in mock repositories are now stored in one MockDatabaseSource. User object now automatically has its role in it instead of a separate Roles table (Role entity was completely removed). Similarly, categories names are now stored in Survey objects, but the Category repository is intact as it is still needed.